requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torchaudio
langchain
langchain-community
langchain-openai
soundfile
numpy